requirements.txt
ollama
faiss-cpu
numpy
pydantic